requirements.txt
Pillow==9.0.1
redis==4.1.4
#torch==1.10.2
